Refactor and enhance type annotations across multiple modules
- Updated type hints for various functions and methods to improve type safety and clarity.
- Introduced `HttpUrl` from Pydantic for URL handling in `core_fetcher.py`, `fetcher.py`, and `crawlers`.
- Enhanced return types in `WebFetcher` and `CachedResourceManager` for better type inference.
- Improved error handling and logging in WebSocket client and manager.
- Added type annotations for dictionaries and lists in parsers and crawlers to ensure consistency.
- Cleaned up unused variables and added type checks to prevent runtime errors.
- Ensured compatibility with static type checkers by adding ignore comments where necessary.

+++ web_fetch/exceptions.py
@@ -420,7 +420,7 @@ class ErrorHandler:
 
         # Exponential backoff for most retryable errors
         if ErrorHandler.is_retryable_error(error):
-            return base_delay * (2**attempt)
+            return float(base_delay * (2**attempt))
 
         return 0.0
